Refactor components for field options to make rendering easier
