+++ ui/js/dfv/src/admin/edit-pod/main-tabs/active-tab-content.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 
 // WordPress dependencies
 import { withSelect, withDispatch } from '@wordpress/data';
