+++ ui/js/dfv/src/admin/edit-pod/main-tabs/field-groups.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import * as PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 import { omit } from 'lodash';
 
 // WordPress dependencies
