+++ ui/js/dfv/src/admin/edit-pod/main-tabs/field-list.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import * as PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { omit } from 'lodash';
