+++ ui/js/dfv/src/admin/edit-pod/main-tabs/settings-modal.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import * as PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 import { __ } from '@wordpress/i18n';
