+++ ui/js/dfv/src/admin/edit-pod/main-tabs/settings-modal.scss
@@ -53,17 +53,8 @@
     }
 }
 
-.pods-form-ui-label {
-    margin-bottom: .2em;
-    display: block;
-}
-
 .pods-field-option label {
     margin-bottom: .2em;
-}
-
-.pods-form-ui-label__required {
-	color: red;
 }
 
 .pods-field-option input[type="text"] {
