+++ ui/js/dfv/src/components/field-description.scss
@@ -1,0 +1,3 @@
+.pods-field-description {
+	clear: both;
+}
