+++ ui/js/dfv/src/components/field-label.scss
@@ -1,0 +1,12 @@
+.pods-field-label {
+    margin-bottom: .2em;
+    display: block;
+}
+
+.pods-field-label__tooltip-wrapper {
+	white-space: nowrap;
+}
+
+.pods-field-label__required {
+	color: red;
+}
