+++ ui/js/dfv/src/components/pods-nav-tab.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 const PodsNavTab = ( { tabs, activeTab, setActiveTab } ) => {
