+++ ui/styles/src/base/_admin.scss
@@ -305,10 +305,6 @@ div.pods-manage-field {
 	zoom: 1;
 }
 
-div.pods-manage-field .description {
-	clear: both;
-}
-
 div.pods-manage-field .pods-field-option,
 div.pods-manage-field .pods-field-option-group {
 	background: #fcfcfc;
@@ -337,7 +333,7 @@ div.pods-manage-field .pods-field-option-group:nth-child(odd) {
 	background: #fff;
 }
 
-div.pods-manage-field .pods-field-option .pods-form-ui-label,
+div.pods-manage-field .pods-field-option .pods-field-label,
 div.pods-manage-field .pods-field-option-group .pods-field-option-group-label {
 	width: 30%;
 	flex: 0 0 30%;
